Split CXF rules into server part and client part
