+++ rule/cxf/src/test/java/io/opentracing/contrib/specialagent/rule/cxf/CxfTest.java
@@ -57,6 +57,8 @@ public class CxfTest {
 
   @Test
   public void testRs(final MockTracer tracer) {
+    System.setProperty(Configuration.INTERCEPTORS_CLIENT_IN, ClientSpanTagInterceptor.class.getName());
+    System.setProperty(Configuration.INTERCEPTORS_SERVER_OUT, ServerSpanTagInterceptor.class.getName());
     final String msg = "hello";
 
     // prepare server
@@ -81,6 +83,11 @@ public class CxfTest {
 
     assertEquals(msg, result);
     assertEquals(4, tracer.finishedSpans().size());
+
+    final List<MockSpan> spans = tracer.finishedSpans();
+    for (final MockSpan span : spans) {
+      assertEquals(AbstractSpanTagInterceptor.SPAN_TAG_VALUE, span.tags().get(AbstractSpanTagInterceptor.SPAN_TAG_KEY));
+    }
 
     server.destroy();
     serverFactory.getBus().shutdown(true);
@@ -107,39 +114,6 @@ public class CxfTest {
 
     assertEquals(msg, response);
     assertEquals(2, tracer.finishedSpans().size());
-
-    server.destroy();
-    serverFactory.getBus().shutdown(true);
-  }
-
-  @Test
-  public void testInterceptor(final MockTracer tracer) {
-    System.setProperty(Configuration.INTERCEPTORS_CLIENT_IN, ClientSpanTagInterceptor.class.getName());
-    System.setProperty(Configuration.INTERCEPTORS_SERVER_OUT, ServerSpanTagInterceptor.class.getName());
-
-    final String msg = "hello";
-
-    // prepare server
-    final JAXRSServerFactoryBean serverFactory = new JAXRSServerFactoryBean();
-    serverFactory.setAddress(BASE_URI);
-    serverFactory.setServiceBean(new EchoImpl());
-    final Server server = serverFactory.create();
-
-    // prepare client
-    final JAXRSClientFactoryBean clientFactory = new JAXRSClientFactoryBean();
-    clientFactory.setServiceClass(Echo.class);
-    clientFactory.setAddress(BASE_URI);
-    final Echo echo = clientFactory.create(Echo.class);
-
-    final String response = echo.echo(msg);
-
-    assertEquals(msg, response);
-    final List<MockSpan> spans = tracer.finishedSpans();
-    assertEquals(2, spans.size());
-
-    for (final MockSpan span : spans) {
-      assertEquals(AbstractSpanTagInterceptor.SPAN_TAG_VALUE, span.tags().get(AbstractSpanTagInterceptor.SPAN_TAG_KEY));
-    }
 
     server.destroy();
     serverFactory.getBus().shutdown(true);
